refactor: move url constants into constants/url.ts
