+++ src/constants/url.ts
@@ -1,0 +1,18 @@
+export const API = {
+  LOGIN: 'https://vinylify-express.vercel.app/',
+  SPOTIFY: 'https://api.spotify.com/v1/',
+};
+export const PAGE = {
+  MAIN: '/',
+  ERROR: '/error',
+  MYPAGE: '/mypage',
+  SEARCH: '/search',
+  MUSIC_INFO: '/music-info',
+  LOGGED_IN: '/me',
+};
+
+export const URL_PARAMS = {
+  KEYWORD: 'keyword',
+  SCOPE: 'scope',
+  PAGE: 'page',
+};
